pyproject.toml
[project]
name = "internettester"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apscheduler>=3.11.0",
    "flask>=3.1.2",
    "pyinstaller>=6.15.0",
    "speedtest-cli>=2.1.3",
    "waitress>=2.1.2",
]
